MToF all monsters
A few things of important note:
There are three files that aren't creatures, but instead describe the traits that can be applied to other monsters. these files are named <thing>template and the 'template' tag has been put in the tags section.

There are also some dwarves that were in the file that I parsed.  As far as I know these aren't officially in any book so I added the tag 'unofficial'.  I didn't want to strip them out because they looked useful.

Lastly, there were some quirks during the parsing and I've gone through and cleaned it up.  It's possible I missed something so not all of the files are guaranteed to be perfect.